+++ NEWS.md
@@ -94,3 +94,4 @@
 * Antigen and serum groups will now be maintained and merged when running `mergeMaps()`
 * Where maps have shared `dilutionStepsize()` settings, this will be applied to the merged map. Alternatively a warning will be issued and a default setting of 1 applied where they have different `dilutionStepsize()` settings.
 * Add attribute getter `numLayers()` for returning the number of titer table layers in a map.
+* `titerTable()<-` will now throw an error if the titer table applied does not have the same dimensions as the map in terms of number of antigens and sera